Refactor slightly how frontend JS is setup
I didn’t want to completely change this, just get what we had working again.

- move polyfills to separate file that is imported
- move `bu_blocks` global object to separate partial `blocks-frontend-tools.js` that can be imported into each frontend.js file so each block can store it’s functions on the single `bu_blocks` object attached to the window.

The polyfills file might not be needed anymore. For one the browser support is better than 2019 so they might be irrelevant polyfills, but also in the old build process there was no browserify or similar performing any JS processing and automatically adding polyfills. Webpack may be doing that as part of the build and that needs to be researched.
