+++ src/blocks-frontend-tools.js
@@ -1,0 +1,4 @@
+// Create a global object to store all BU Blocks related functions and data.
+const bu_blocks = {};
+
+export default bu_blocks;
